Add new robust Synology backup bash script
A backup script written in bash that uses rsync for transfers.

Features:

A locking mechanism.

Traps for SIGHUP, SIGINT, SIGQUIT, SIGTERM signals, and EXIT, which
cause different things to happen depending on the function they were
caught in.

Neatly organized logging for each share that is to be backed up; each
log file is sed-filtered to remove files excluded by rsync.

Plaintext emails describing the results of transfers, lock failures, and
caught signals.

Caveats:

Email is currently sent via a physically separate server that has an
SMTP client which supports OAuth2 (msmtp).

This script is not POSIX-compliant and likely never will be.

This is the public version which has been sanitized.