Add feature to handle Friday and Sunday promos
post_enc_promo.bash will now handle the case on Fridays where there is a
weekend DP in addition to the regular DP scheduled for that day,
including sending a slightly different email.

A similar feature has also been added to handle the case where there is
a weekend social promo in addition to the daily social promo.