Add a backup script for *nix machines
server_backups.bash will, given a list of server:ssh_port key-value
pairs and remote directories, back them up to the server the script is
run on.

Assumes that rsync is installed on the remote hosts and can be run as
root, and that passwordless ssh is set up. An ansible playbook can help
with this initial provisioning.