Add small script used to delete weekly AWS files
This script is used to delete the weekly media files uploaded to a
particular AWS S3 bucket, and the subsequent video files spit out by
MediaConvert jobs.

Caveat: The AWS CLI returns an exit code of 0 even if there are no files
to delete, so be sure to use the correct show number.