Add FFmpeg example for multiple audio stream files
For each video file, the editors also export 1-3 .aiff audio files, for
the dialogue, effcts, and music. These are mixed into one audio stream,
and added to the final mezzanine video file that they export.

They may want to start providing video files that have all the audio
streams in the MOV container, and expect the encoder to do the mixing.

This multitrack.bash file contains an example of how to do this.